requirements.txt
aiogram==3.13.1
Pillow==10.4.0
pypdf==4.3.1
pdfplumber==0.11.4
python-docx==1.1.2
PyMuPDF==1.24.11
python-dotenv==1.0.1
aiohttp==3.10.5
pytesseract==0.3.10
